+++ mopidy/stream/actor.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, unicode_literals
 import fnmatch
 import logging
 import re
+import time
 import urlparse
 
 import pykka
@@ -80,22 +81,70 @@ class StreamPlaybackProvider(backend.PlaybackProvider):
                 stream.Extension.dist_name, stream.Extension.version))
 
     def translate_uri(self, uri):
-        try:
-            scan_result = self._scanner.scan(uri)
-        except exceptions.ScannerError as e:
-            logger.warning(
-                'Problem scanning URI %s: %s', uri, e)
-            return None
+        return unwrap_stream(
+            uri,
+            timeout=self._config['stream']['timeout'],
+            scanner=self._scanner,
+            requests_session=self._session)
 
-        if not (scan_result.mime.startswith('text/') or
+
+def unwrap_stream(uri, timeout, scanner, requests_session):
+    """
+    Get a stream URI from a playlist URI, ``uri``.
+
+    Unwraps nested playlists until something that's not a playlist is found or
+    the ``timeout`` is reached.
+    """
+
+    original_uri = uri
+    deadline = time.time() + timeout
+
+    while time.time() < deadline:
+        logger.debug('Unwrapping stream from URI: %s', uri)
+
+        try:
+            scan_timeout = deadline - time.time()
+            if scan_timeout < 0:
+                logger.info(
+                    'Unwrapping stream from URI (%s) failed: '
+                    'timed out in %sms',
+                    uri, timeout)
+                return None
+            scan_result = scanner.scan(uri, timeout=scan_timeout)
+        except exceptions.ScannerError as exc:
+            logger.debug('GStreamer failed scanning URI (%s): %s', uri, exc)
+            scan_result = None
+
+        if scan_result is not None and not (
+                scan_result.mime.startswith('text/') or
                 scan_result.mime.startswith('application/')):
+            logger.debug(
+                'Unwrapped potential %s stream: %s', scan_result.mime, uri)
             return uri
 
-        content = http.download(self._session, uri)
+        download_timeout = deadline - time.time()
+        if download_timeout < 0:
+            logger.info(
+                'Unwrapping stream from URI (%s) failed: timed out in %sms',
+                uri, timeout)
+            return None
+        content = http.download(
+            requests_session, uri, timeout=download_timeout)
+
         if content is None:
+            logger.info(
+                'Unwrapping stream from URI (%s) failed: '
+                'error downloading URI %s', original_uri, uri)
             return None
 
-        tracks = list(playlists.parse(content))
-        if tracks:
-            # TODO Test streams and return first that seems to be playable
-            return tracks[0]
+        uris = playlists.parse(content)
+        if not uris:
+            logger.debug(
+                'Failed parsing URI (%s) as playlist; found potential stream.',
+                uri)
+            return uri
+
+        # TODO Test streams and return first that seems to be playable
+        logger.debug(
+            'Parsed playlist (%s) and found new URI: %s', uri, uris[0])
+        uri = uris[0]